backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
pydantic==1.10.13
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
reportlab==4.0.4
matplotlib==3.8.1
scikit-learn==1.3.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
requests==2.31.0